Fix bug with large arrays
If you provide a large enough array of numbers (100K is pretty
consistent) then you're likely to get Integer.MIN_VALUE as the
result. This seems impossible since you're returning the absolute
value of the difference of a pair of values. It turns out that this
is because a weird issue with Java and integer representation.

Because of the asymmetry caused by the presence of 0 in the int
type, Integer.MIN_VALUE (-2147483648) has magnitude one greater
than Integer.MAX_VALUE (+2147483647). So if the difference of
the two values is Integer.MIN_VALUE, then attempting to take
its absolute value leads to a number that can't be represented
as an int and you end up (due to the details of two's complement)
flipping back to Integer.MIN_VALUE!

Changing the return value from `int` to `long` should take
care of this.

+++ src/mpd/SerialMain.java
@@ -14,7 +14,7 @@ public class SerialMain {
             values[i] = random.nextInt();
         }
         
-        int result = mpd.minimumPairwiseDistance(values);
+        long result = mpd.minimumPairwiseDistance(values);
         System.out.println("The minimum distance was " + result);
     }
